pyproject.toml
[project]
name = "pulseiq"
version = "0.1.0"
description = "Data engineering centric AI chatbot project"
requires-python = ">=3.11"
dependencies = [
  "dbt-duckdb>=1.9.0",
  "duckdb>=1.1.0",
  "faker>=37.0.0",
  "fastapi>=0.115.0",
  "openai>=1.75.0",
  "pandas>=2.2.0",
  "pydantic>=2.9.0",
  "python-dotenv>=1.0.1",
  "pyarrow>=18.0.0",
  "qdrant-client>=1.9.0",
  "uvicorn>=0.30.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
